+++ src/main/java/faang/school/postservice/dto/comment/CommentEventDto.java
@@ -1,0 +1,16 @@
+package faang.school.postservice.dto.comment;
+
+import lombok.*;
+
+@Getter
+@Setter
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class CommentEventDto {
+    private long userId;
+    private long authorId;
+    private long postId;
+    private String content;
+    private long commentId;
+}
