+++ src/main/java/faang/school/postservice/publisher/MessagePublisher.java
@@ -1,0 +1,5 @@
+package faang.school.postservice.publisher;
+
+public interface MessagePublisher<T> {
+    void publish(T message);
+}
