+++ src/test/java/faang/school/postservice/service/moderator/ModeratorServiceTest.java
@@ -2,6 +2,8 @@ package faang.school.postservice.service.moderator;
 
 import faang.school.postservice.config.dictionary.OffensiveWordsDictionary;
 import faang.school.postservice.model.Comment;
+import faang.school.postservice.model.Post;
+import faang.school.postservice.publisher.PublishedCommentEventPublisher;
 import faang.school.postservice.service.comment.CommentService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
@@ -15,13 +17,21 @@ import java.util.List;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 
+import static java.util.concurrent.TimeUnit.SECONDS;
+import static org.awaitility.Awaitility.await;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.mockito.Mockito.*;
 
 @ExtendWith(MockitoExtension.class)
 class ModeratorServiceTest {
 
+    private static final int TWO_TIMES_USED = 2;
     private static final int FOUR_TIMES_USED = 4;
+
+    private static final long COMMENT_ID = 1L;
+    private static final long POST_ID = 1L;
+    private static final long USER_ID = 1L;
+    private static final long FIVE_SECOND_AWAIT = 5L;
 
     @Mock
     private CommentService commentService;
@@ -29,33 +39,50 @@ class ModeratorServiceTest {
     @Mock
     private OffensiveWordsDictionary offensiveWordsDictionary;
 
+    @Mock
+    private PublishedCommentEventPublisher publishedCommentEventPublisher;
+
     private ModeratorService moderatorService;
 
     private final ExecutorService executorService = Executors.newSingleThreadExecutor();
 
+    private Post post;
     private Comment comment;
     private Comment comment1;
     private List<Comment> unverifiedComments;
 
     @BeforeEach
     void init() {
-        moderatorService = new ModeratorService(commentService, executorService, offensiveWordsDictionary);
+        moderatorService = new ModeratorService(commentService,
+                executorService,
+                offensiveWordsDictionary,
+                publishedCommentEventPublisher);
 
-    }
+        post = Post.builder()
+                .id(POST_ID)
+                .authorId(USER_ID)
+                .build();
 
-
-    @Test
-    @DisplayName("When no offensive words then update comment and save")
-    void whenNoOffensiveWordsThenUpdateAndSaveComments() {
         comment = Comment.builder()
+                .id(COMMENT_ID)
+                .authorId(USER_ID)
+                .post(post)
                 .content("test test1")
                 .build();
 
         comment1 = Comment.builder()
+                .id(COMMENT_ID)
+                .authorId(USER_ID)
+                .post(post)
                 .content("a b")
                 .build();
 
         unverifiedComments = Arrays.asList(comment, comment1);
+    }
+
+    @Test
+    @DisplayName("When no offensive words then update comment and save")
+    void whenNoOffensiveWordsThenUpdateAndSaveComments() {
         when(commentService.getUnverifiedComments())
                 .thenReturn(unverifiedComments);
         when(offensiveWordsDictionary.isWordContainsInDictionary(anyString()))
@@ -64,26 +91,22 @@ class ModeratorServiceTest {
         moderatorService.moderateCommentsContent();
 
         verify(commentService).getUnverifiedComments();
-         verify(offensiveWordsDictionary, times(FOUR_TIMES_USED)).isWordContainsInDictionary(anyString());
         verify(commentService).saveComments(anyList());
 
         assertEquals(Boolean.TRUE, comment.isVerified());
         assertEquals(Boolean.TRUE, comment1.isVerified());
+
+        await().atMost(FIVE_SECOND_AWAIT, SECONDS).untilAsserted(() -> {
+
+            verify(offensiveWordsDictionary, times(FOUR_TIMES_USED)).isWordContainsInDictionary(anyString());
+
+            verify(publishedCommentEventPublisher, times(TWO_TIMES_USED)).publish(any());
+        });
     }
 
-
-   @Test
+    @Test
     @DisplayName("When offensive words contains then update comment and save")
     void whenOffensiveWordsThenUpdateAndSaveComments() {
-        comment = Comment.builder()
-                .content("test test1")
-                .build();
-
-        comment1 = Comment.builder()
-                .content("a b")
-                .build();
-
-        unverifiedComments = Arrays.asList(comment, comment1);
         when(commentService.getUnverifiedComments())
                 .thenReturn(unverifiedComments);
         when(offensiveWordsDictionary.isWordContainsInDictionary(anyString()))
@@ -92,10 +115,14 @@ class ModeratorServiceTest {
         moderatorService.moderateCommentsContent();
 
         verify(commentService).getUnverifiedComments();
-        verify(offensiveWordsDictionary, times(FOUR_TIMES_USED)).isWordContainsInDictionary(anyString());
         verify(commentService).saveComments(anyList());
 
         assertEquals(Boolean.FALSE, comment.isVerified());
         assertEquals(Boolean.FALSE, comment1.isVerified());
+
+        await().atMost(FIVE_SECOND_AWAIT, SECONDS).untilAsserted(() -> {
+            verify(offensiveWordsDictionary, times(FOUR_TIMES_USED)).isWordContainsInDictionary(anyString());
+            verify(publishedCommentEventPublisher, never()).publish(any());
+        });
     }
 }
